Turn watch command into a flag for test command
The watch command resulted in some duplicated code and didn't need to be
a command on its own. Having it be a flag on the test command is enough.
