+++ src/cli.rs
@@ -31,32 +31,10 @@ pub fn init() -> App<'static, 'static> {
                     .arg(Arg::with_name("fetch")
                          .long("fetch")
                          .help("If the test folder does not exist, download the test files from Kattis"))
-                   )
-          // FIXME: Very duplicated from the above test command
-          .subcommand(SubCommand::with_name("watch")
-                    .about("Runs a solution through the official test cases every time the source file changes. Takes the same arguments as the test command.")
-                    .setting(AppSettings::DisableVersion)
-                    .arg(Arg::with_name("PATH")
-                         .help("Path to problem directory")
-                         .default_value(".")
-                         .index(1))
-                    .arg(Arg::with_name("file")
-                         .short("f")
-                         .long("file")
-                         .takes_value(true)
-                         .help("Name of source file to test. Necessary when there are multiple valid sources or when the program cannot recognise the file extension"))
-                    .arg(Arg::with_name("language")
-                         .short("l")
-                         .long("lang")
-                         .takes_value(true)
-                         .help("Programming language of the solution. Write the typical file extension for the language (java for Java, py for python, js for JavaScript, etc.)."))
-                    .arg(Arg::with_name("time")
-                         .short("t")
-                         .long("time")
-                         .help("Display how long each test case took to execute"))
-                    .arg(Arg::with_name("fetch")
-                         .long("fetch")
-                         .help("If the test folder does not exist, download the test files from Kattis"))
+                    .arg(Arg::with_name("watch")
+                         .short("w")
+                         .long("watch")
+                         .help("Re-runs tests every time the source file changes"))
                    )
         .subcommand(SubCommand::with_name("get")
                     .about("Fetches a problem from Kattis by creating a directory of the same name and downloading the official test cases")
