+++ src/commands/mod.rs
@@ -7,7 +7,6 @@ mod random;
 mod submit;
 mod test;
 mod update;
-mod watch;
 
 pub use config::config;
 pub use get::get;
@@ -18,4 +17,3 @@ pub use random::random;
 pub use submit::submit;
 pub use test::test;
 pub use update::update;
-pub use watch::watch;
